requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-core==2.16.1
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.8
python-multipart==0.0.6
python-dotenv==1.2.1
